Agregados controledores necesarios

+++ src/main/java/com/proyectotingeso/sistemasueldos/repositorios/EmpleadoRepositorio.java
@@ -3,5 +3,5 @@ package com.proyectotingeso.sistemasueldos.repositorios;
 import com.proyectotingeso.sistemasueldos.entidades.EmpleadoEntidad;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EmpleadoRepositorio extends CrudRepository <EmpleadoEntidad, Long>{
+public interface EmpleadoRepositorio extends CrudRepository <EmpleadoEntidad, Integer>{
 }
